requirements.txt
awesome-slugify
PyYAML
requests

